lab_01/requirements.txt
kafka-python
pandas
Path
numpy